Bug fixes and algorithm improvement
-- Fixed bug which made the second AI player start in the right
direction. This would lead the AI to believe the next move was [1, 0]
from [0,0]. This is illegal by game rules since the next move following
the initial start of the game is always [0, 1].

-- This bug also caused the AI to report the wrong coordinates. More
specifically, the coordinates were flipped, i.e.. [ y, x ] instead of
[x, y ]

-- The fix involved initializing the current direction to DOWN in the
constructor. That way, every UVI player starts in the default direction.